feature: add cli
