+++ main_test.py
@@ -1,4 +1,4 @@
-import generateJSON
+import main
 from domain.Room import Room
 from domain.Geometry import Geometry
 from domain.Point3D import Point3D
@@ -27,7 +27,7 @@ def test_get_coord():
 
     coordinates = {"X": [0, 1, 1, 0], "Y": [0, 0, 1, 1]}
 
-    assert generateJSON.get_coord(building_element) == coordinates
+    assert main.get_coord(building_element) == coordinates
 
 
 def test_point_inside_edge():
@@ -36,6 +36,6 @@ def test_point_inside_edge():
     elem2_points_x = [0.5, 1, 1, 0.5]
     elem2_points_y = [-0.5, -0.5, 0, 0]
 
-    assert generateJSON.is_in_edge(
+    assert main.is_in_edge(
         x=elem2_points_x, y=elem2_points_y, xp=elem1_points_x, yp=elem1_points_y
     )
